Avoid recreation of client object for each index access
See https://github.com/hbz/lobid-resources/issues/286
